Chore: move some packages to the new internal dir
