+++ internal/interfaces/interfaces.go
